ncp: Source cleanup and refactor.
This change gets rid of the huge switch statements and now
uses jump tables to look up the appropriate method to call for
a given command-id or property. The result is much cleaner and,
hopefully, easier to maintain and extend.

Eventually we may want to numerically sort the lists so that
we can use a binary search instead of a linear search, but
for now this will do.