README: Improve markdown formatting for plaintext readability
This change doesn't affect how the README looks when viewed rendered
into HTML. It simply reformats the markdown in such a way that
improves readability when reading the document using a plaintext
editor.

One of the ways to improve the visual appeal of the text is to use
reference-style links whenever an inline-link would cause the
paragraph to line-wrap in a visually unappealing way.

For example, this is before:

    Submit bugs and feature requests to [issue
    tracker](https://github.com/openthread/openthread/issues). Usage
    questions? Post questions to [Stack
    Overflow](http://stackoverflow.com/) using the [`openthread`
    tag](http://stackoverflow.com/questions/tagged/openthread). We also
    use Google Groups for discussion and announcements:

And this is after:

    Submit bugs and feature requests to [issue tracker][ot-issues]. Usage
    questions? Post questions to [Stack Overflow][stackoverflow] using the
    [`openthread` tag][ot-tag]. We also use Google Groups for discussion
    and announcements:

    [ot-issues]: https://github.com/openthread/openthread/issues
    [stackoverflow]: http://stackoverflow.com/
    [ot-tag]: http://stackoverflow.com/questions/tagged/openthread