key_manager: Don't set changed flag in key sequence getter
The previous code was in error because the key sequence changed
flag tends to induce the key sequence to be read again, which
causes the flag to be set again, and so on.

This change moves the key sequence changed flag to be set
only in places where the key sequence is actually being changed.