NcpBase: Fix the model for adding/removing IPv6 addresses. (#219)
This change addresses issue #204 by adding an array of elements
of type `otNetifAddress` to `NcpBase` class. The array elements
are then used for adding/removing of Ipv6 addresses. It also
modifies the logic for adding an already added address. In such a
case the prefix length and address lifetime values for the address
get updated and operation is considered successful.