Adding new spinel property to allow net data changes and registering local data with leader
This commit adds a new spinel prop 'THREAD_ALLOW_LOCAL_NET_DATA_CHANGE`
with boolean type. This property should be set to `true` before
modifying any of local network data propeties (e.g., inserting a
new prefix to `ON_MESH_NETS` or to `LOCAL_ROUTES`).

When this property is changed from `true` to `false`, any changed
to the local network data is registered with the leader.

This commit is related to issue #272.