ncp-uart: Keep track of and report framing errors. (#371)
Previously we were simply ignoring frames that contained errors. This
change causes framing errors to be logged to the `PROP_STREAM_DEBUG`
stream property, so that the event can be logged by the host. The log
message looks like this:

    Framing error E: [XX XX XX XX...]

Where `E` is the `ThreadError` code, and `XX XX...` is a hex dump of
no more than the first 42 bytes of the erroneous frame.

A frame error counter is also included, but is not yet being exposed.