ncp: Adds new property, 'PROP_NET_REQUIRE_JOIN_EXISTING' (#497)
This flag is typically used for nodes that are associating with an
existing network for the first time. If this is set to `true` before
`PROP_NET_STACK_UP` is set to `true`, the creation of a new partition
at association is prevented. If the node cannot associate with an
existing partition, `PROP_LAST_STATUS` will emit a status that
indicates why the association failed and `PROP_NET_STACK_UP` will
automatically revert to `false`.

Once associated with an existing partition, this flag automatically
reverts to `false`.