Ncp: Change the encoding of child info in prop THREAD_CHILD_TABLE getter
This commit makes two changes in how the child info is encoded in
the getter of property `THREAD_CHILD_TABLE`. First, the child mode
flags (e.g., `RxOnWhenIdle`) are sent as a single uint8_t bitmask.
Second, the child id is no longer included in the response (as it
can be derived from Rloc16).