Refinement of Spinel Protocol Documentation
This commit refines the documentation of the Spinel NCP protocol. What
was previously a single markdown-based document has now been broken
into multiple files which can be edited independently. These files can
then be compiled into a final document in either XML, HTML, or
plaintext formats. The [`mmark`](https://github.com/miekg/mmark) tool is
used for compiling the source documentation into a single XML file and
[`xml2rfc`](https://xml2rfc.tools.ietf.org/) is used to convert from
the XML to the final desired format.

Note that the building of the protocol documents is handled
independently from the normal OpenThread build. The spinel protocol
document source files are intentionally excluded from the build
system, since there is no value in including them. Copies of the final
compiled documents are included to ensure that they are distributed
with any release archives.

This commit is focused on introducing this new strucutre to the protocol
document. The document itself is a work in progress and may contain
errors or require revisions, hence the `draft-` prefix in the filenames.

This commit is related to issue #492, "☂ Spinel NCP Protocol
Standardization".