Add implementations for strlcpy/strlcat when missing
This commit shows how to add implementations of missing functions for
platforms where those functions are not present. In this case we are
adding the `strlcpy()` and `strlcat()`, which are "safe" replacements
for `strncpy()` and `strncat()` that behave in a way more consistent
with developer expectations.

After this commit, adding additional replacements for missing
functions should be fairly straightforward.

[1]: http://stackoverflow.com/questions/3821916/how-to-merge-two-ar-static-libraries-into-one
[2]: https://www.gnu.org/software/automake/manual/html_node/Libtool-Convenience-Libraries.html