spi-hdlc-adapter: Improved strategy for how SPI transactions are performed.
The big change in this commit is a change in the transmission strategy used.
This change isn't fixing any particular bug, but these changes should improve
performance.

Instead of always using a throw-away transaction to figure out the buffer sizes
for the NCP, we now attempt to send any frame we have in the first transaction.
If the NCP has a response that happens to fit in the size of that packet, it
will be processed.

This change also adds a lot more comments better explaining the more
complicated parts of the code.