NcpSpi: Fix a race situation and simply the tx state model
This commit fixes a race issue in tx path of `NcpSpi`: If the SPI
transaction is completed while in middle of `PrepareTxFrame()`task
handler before clearing the `mHandlingSendDone` flag, we can
have a situation where `PrepareTxFrame` task is not posted again and
then the next queued tx frames are not handled immediately.

It also simplifies the tx state model by adding a single `mTxState`
variable to track the current state as an `enum`.