Add new macro `IgnoreReturnValue()`
This commit adds a new macro `IgnoreReturnValue(statement)` which
runs the `statement` (which should be a function/method which
returns a status/value) and ignores the returned value.

This macro is primarily used to indicate the intention of developer
that the return value of the function/method can be safely ignored.