NcpBase: Separate sending of network data from list of on-mesh prefixes (#1219)
This commit separates the sending of full network data and the list
of on-mesh prefixes when an `OT_THREAD_NETDATA_UPDATED` event is
received. The list of on-mesh prefixes is sent out only after the
network data is sent out (or queued) successfully. This handles
the case where the NCP tx buffer may not have enough space to fit
both messages.