Simplify the handling of unicast/multicast address arrays
Change the logic in `Netif` to remove the use of `uint8_t` bitmask
to indicate if an address entry in the array is in-use. Instead to
mark the unused entries, set the `mNext` pointer to point back to
the entry itself. This ensures that `MAX_EXT_IP_ADDRS` can be safely
set to a value higher than 8.