`mle`: Set `mRetrieveNewNetworkData` to `true` when requesting network data
This change ensures that the flag `mRetrieveNewNetworkData` is set to
`true` when a "Data Request" message is sent.

This addresses an issue where after updating the `mLeaderData`
(e.g., from a "ChildUpdateResponse" message not containing the active
dataset), and requesting (from `SendDataRequest()`) and receiving the
network data, `mRetrieveNewNetworkData` could remain `false` and
cause the network data not to be updated locally.