Let platforms specify linker section for version string
`PLATFORM_VERSION_ATTR_PREFIX` and `PLATFORM_VERSION_ATTR_SUFFIX` are
intended to be used to specify compiler directives to indicate
what linker section the platform version string should be stored.

This is useful for specifying an exact locaiton of where the version
string will be located so that it can be easily retrieved from the
raw firmware image.

If `PLATFORM_VERSION_ATTR_PREFIX` is unspecified, the keyword `static`
is used instead.

If both are unspecified, the location of the string in the firmware
image will be undefined and may change.