spi-hdlc-adapter: Add support for raw mode. (#1341)
This commit adds support for two new arguments:

*   `--raw`: Do not encode/decode packets using HDLC. Instead, write
    whole, raw frames to the specified input and output FDs. This is
    useful for emulating a serial port, or when datagram-based sockets
    are supplied for `stdin` and `stdout` (when used with `--stdio`).
*   `--mtu=[MTU]`: Specify the MTU. Currently only used in raw mode.
    Default and maximum value is 2043. Must be greater than zero.

The `--raw` argument will allow `wpantund` to use `spi-hdlc-adapter`
without HDLC escaping by supplying datagram sockets for stdin and
stdout. This should improve performance and reduce latency.

The `--mtu` argument is supplied to benefit the use case where we
are emulating an asynchronous serial port over SPI. In such a case,
you would likely want to lower the MTU significantly to match the
internal buffer size of the MCU.

Version is bumped to 0.04.