`NCP/Spinel` Add new spinel property to test assert
This commit adds `SPINEL_PROP_DEBUG_TEST_ASSERT` as a new read-only
spinel property and also adds its associated get handler.

Getting this property will cause an assert on the NCP. This is
intended as a way for testing the the assert functionality of
underlying platform.

The spinel proocol documents (`md` files) are also updated
accordingly.