spi-hdlc-adapter: Debugging and minor performance improvement. (#1468)
This change primarily adds additional debug logging and statistics
counters. It also includes the ability to change the small-packet
size, which was previously hard-coded to a value of 5. The default
value for the small-packet size has been increased to 32, which
improves performance.

We also now do a better job of tracking flow control and slave
liveliness.