`Timer`: Update how we detect if a timer is running
This commit makes some changes in the `Timer` implementation to
simplify how we detect if a timer is running.

On an unscheduled timer, we ensure to set the `mNext` to point back
to the `Timer` instance itself. In a scheduled/running timer, the
timer is included in a linked-list where `mNext` is used to point
to next timer in the list (or `NULL` if at the end of list). The
`mNext` value is then used to quickly determine if a timer is
scheduled/running or not.

This commit also removes the now unused `TimerScheduler::IsAdded()`
method.