CLI fix for printing child list
In the previous implementation, if a child was not available at a
certain index, all children at greater indexes would not be
printed for a child list or child table command, despite still being
connected to the node.

This change modifies the logic so that if a child does not exist at
a given index, only that index is disgarded, and future indexes are
checked correctly.