`NcpSpi`: Enable full duplex SPI exchange.
This commit implements support for full duplex SPI exchange in
`NcpSpi`. To realize this a slight modification is made to behavior of
`otPlatSpiSlavePrepareTransaction()` under `spi-slave.h` to allow a
`NULL` pointer to be passed in as either `aOutputBuf` or `aInputBuf`
input arguments and for it to mean that that buffer pointer should not
change from its previous/current value. This allows output buffer to
be changed safely from a tasklet when there is new outbound frame without
changing the input buffer pointers.