Implement "last-in wins" duplicate IPv6 address resolution (#1720)
This commit implements "last-in wins" duplicate IPv6 address
resolution policy for children's registered addresses. When a
child registers a new address, it is checked if the same IPv6
address is registered by another child and if so the address is
removed from the other child's registered IPv6 address list.

Duplicate addresses can occur if a previously attached child
attaches to same parent again (after a reset, memory wipe) using
a new random extended address before the old entry in the child
table is timed out and then trying to register its globally unique
IPv6 address as the new child.