coap API: Add Set Content Format API call. (#1889)
* coap API: Add Append Content Format API call.

This allows us to easily set the content format of outgoing messages by
wrapping otCoapHeaderAppendUintOption.  Additionally, some standard CoAP
content format codes are defined as an enum.

[Supercedes commit 52cddd3a4f4127dc22e64b65fd9392dc09aa79d8]

* coap header: Re-work AppendContentFormatOption

Remove the practically empty `MediaType` enum and replace it with the
more complete (but more verbose) `otCoapOptionContentFormat` so that it
is consistent with other CoAP API functions.

* coap header: Use C++ static cast

As per review notes (pull request #1889).

* coap api: Add blank line between @param and @retval

As per review comments (pull request #1889).