[mesh-forwarder] Drop a message if all tx attempts fail for one of its fragments
Existing code already enables this behavior for indirect message
transmissions. This commits adds support for direct transmission and
also introduces `OPENTHREAD_CONFIG_DROP_MESSAGE_ON_FRAGMENT_TX_FAILURE`
as a new configuration option to control this feature. If enabled
OpenThread will drop a message (and not send any remaining fragments
of the message) if all transmit attempts fail for a fragment of the
message. For a direct transmission, a failure occurs after all MAC
transmission attempts for a given fragment are unsuccessful. For an
indirect transmission, a failure occurs after all data poll triggered
transmission attempts for a given fragment fail. If this feature is
disabled, OpenThread will attempt to send subsequent fragments,
whether or not all transmission attempts fail for a given fragment.
Default setting enables this behavior.