[mesh-forwarder] remove `BUSY` error log in `ScheduleTransmissionTask()`
This commit removes the `error` from `ScheduleTransmissionTask()` and
also removes the corresponding log line printing the error. The BUSY
error log is not really helpful as `ScheduleTransmissionTask` is
posted whenever a new messages is queued and within this method it is
first checked if we are busy (i.e., in middle of an earlier tx) and
simply return and wait for end of current tx to process the next one.