[mac] Change tx state/operation tracking
This commit changes how the MAC layer tracks the current (tx related)
operation (e.g., active/energy scan, beacon transmission or data
transmission). The operation can be changed from `StartOperation()`
and when it is and radio is put to idle state from `EndOperation()`
(e.g, when scan finished). `StartOperation()` ensures that next
operation is not scheduled until the current one is finished.